legal: add license
I, Tom Oostveen, as sole contributor to my own dotfiles, allow anyone to
make use or copy everything in this repository under the terms of the
MIT license, including past contributions.

I just forgot to add the MIT license in the first place that's all.

I figured it would good to include this so anyone can steal parts of my
config if they want (though you might not actually want to do that)